+++ src/classes/babyClasses.ts
@@ -228,6 +228,7 @@ export { PsychicBaby } from "./babies/PsychicBaby";
 export { PubicBaby } from "./babies/PubicBaby";
 export { PuffBaby } from "./babies/PuffBaby";
 export { PuncherBaby } from "./babies/PuncherBaby";
+export { PunkboyBaby } from "./babies/PunkboyBaby";
 export { PurpleBaby } from "./babies/PurpleBaby";
 export { PuzzleBaby } from "./babies/PuzzleBaby";
 export { RabbitBaby } from "./babies/RabbitBaby";